% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smith}
\alias{smith}
\title{Smith 1907 Cortical Atlas}
\usage{
smith()
}
\value{
A [ggseg.formats::ggseg_atlas] object (cortical).
}
\description{
A historical Smith 1907 atlas based on files released with
Pijnenburg et al., NeuroImage, 239, 2021.
}
\examples{
\dontrun{
smith()
}
}
\references{
G.E. Smith; A new topographical survey of the human cerebral cortex,
being an account of the distribution of the anatomically distinct
cortical areas and their relationship to the cerebral sulci;
J. Anat. Physiol., 41 (Pt 4) (1907), p. 237

Pijnenburg R, et al. (2021). \doi{10.1016/j.neuroimage.2021.118274}
}
\concept{ggseg_atlases}
